requirements.txt
appdirs==1.4.4
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cssselect==1.1.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.2
django-countries==7.2.1
django-crispy-forms==1.14.0
django-environ==0.8.1
django-etc==1.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fake-useragent==0.1.11
filelock==3.6.0
geographiclib==1.52
gunicorn==20.1.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
itemadapter==0.5.0
itemloaders==1.0.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
lxml==4.8.0
MarkupSafe==2.1.1
oauthlib==3.2.0
outcome==1.2.0
parse==1.19.0
parsel==1.6.0
Pillow==9.0.1
Protego==0.2.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyee==8.2.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
selenium==4.4.3
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tldextract==3.2.0
tqdm==4.63.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
validators==0.20.0
w3lib==1.22.0
websockets==10.2
whitenoise==6.0.0
wsproto==1.2.0
zipp==3.7.0
zope.interface==5.4.0
